+++ banner-tweaks/common/src/main/java/com/chimericdream/bannertweaks/config/BannerTweaksConfig.java
@@ -1,35 +1,20 @@
 package com.chimericdream.bannertweaks.config;
 
 import com.chimericdream.bannertweaks.ModInfo;
-import com.google.gson.GsonBuilder;
+import com.chimericdream.lib.config.YaclConfig;
 import dev.isxander.yacl3.api.ConfigCategory;
 import dev.isxander.yacl3.api.Option;
 import dev.isxander.yacl3.api.OptionDescription;
-import dev.isxander.yacl3.api.YetAnotherConfigLib;
 import dev.isxander.yacl3.api.controller.IntegerSliderControllerBuilder;
-import dev.isxander.yacl3.config.v2.api.ConfigClassHandler;
 import dev.isxander.yacl3.config.v2.api.SerialEntry;
-import dev.isxander.yacl3.config.v2.api.serializer.GsonConfigSerializerBuilder;
-import dev.isxander.yacl3.platform.YACLPlatform;
-import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.Identifier;
 
 public class BannerTweaksConfig {
     @SerialEntry
     public int maxBannerLayers = Defaults.MAX_BANNER_LAYERS;
 
-    public static ConfigClassHandler<BannerTweaksConfig> HANDLER = ConfigClassHandler.createBuilder(BannerTweaksConfig.class)
-        .id(Identifier.fromNamespaceAndPath(ModInfo.MOD_ID, "config"))
-        .serializer(config -> GsonConfigSerializerBuilder.create(config)
-            .setPath(YACLPlatform.getConfigDir().resolve("bannertweaks.json5"))
-            .appendGsonBuilder(GsonBuilder::setPrettyPrinting)
-            .setJson5(true)
-            .build())
-        .build();
-
-    public static Screen configScreen(Screen parent) {
-        return YetAnotherConfigLib.create(HANDLER, ((defaults, config, builder) -> builder
+    public static final YaclConfig<BannerTweaksConfig> CONFIG = YaclConfig.builder(BannerTweaksConfig.class, ModInfo.MOD_ID)
+        .screen((defaults, config, builder) -> builder
             .title(Component.literal("Banner Tweaks Config"))
             .category(ConfigCategory.createBuilder()
                 .name(Component.literal("Banner Tweaks Config"))
@@ -42,8 +27,8 @@ public class BannerTweaksConfig {
                         .step(1))
                     .build())
                 .build())
-        )).generateScreen(parent);
-    }
+        )
+        .build();
 
     public static class Defaults {
         public static int MAX_BANNER_LAYERS = 12;