+++ banner-tweaks/common/src/main/java/com/chimericdream/bannertweaks/mixin/BannerPatternsComponentMixin.java
@@ -10,6 +10,6 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 public class BannerPatternsComponentMixin {
     @ModifyConstant(method = "addToTooltip", constant = @Constant(intValue = 6))
     private int getLimit(int curr) {
-        return BannerTweaksConfig.HANDLER.instance().maxBannerLayers;
+        return BannerTweaksConfig.CONFIG.instance().maxBannerLayers;
     }
 }