+++ banner-tweaks/common/src/main/java/com/chimericdream/bannertweaks/mixin/LoomMenuMixin.java
@@ -10,6 +10,6 @@ import org.spongepowered.asm.mixin.injection.ModifyConstant;
 public class LoomMenuMixin {
     @ModifyConstant(method = "containerChanged", constant = @Constant(intValue = 6))
     public int getLimit(int constant) {
-        return BannerTweaksConfig.HANDLER.instance().maxBannerLayers;
+        return BannerTweaksConfig.CONFIG.instance().maxBannerLayers;
     }
 }